[!]Rip off defaults MelOpt Float - change of behavior:
EEE

super(MelRefScale, self).__init__(b'XSCL', u'ref_scale', 1.0)

As discussed having "optional" elements turns out to be not such a good
idea:
- actually *most* of the elements should be optional. In fact this commit
works in this direction as flipping the default to None for _MelNum
(by deleting __init__ override and modifying pack_subrecord_data) makes
all those numbers optional. Note MelFid was already there. Might break
if we relied on 0 default for creating a record - we don't do this
much but still !
- so we should make some elements required (and supply defaults for
these) - this is quite some effort and seems that nobody is sure which
are those required elements - so better left for later (should be
handled at Mel level - so after having a mel hierarchy)
- some of the "optional" elements specified non zero (non None now)
defaults. Those would not be dumped if we did indeed load that value
and this is a change that might break stuff, if we relied on this to
drop records for some reason - but we didn't see:

https://github.com/wrye-bash/wrye-bash/commit/0ec45db55d6a117105b0b433613c4a2160831ac0#commitcomment-47085582

- Mopy/bash/game/oblivion/patcher/special.py: default 0 was needed here
- the cases that we set the default to null are handled naturally

Net change:
- We don't omit the (few) elements if present and have 0 (or the handful
of non 0 MelFloats). We don't want to do this
- MelFloat (_Num later) won't be dumped if not present - we want this